Remove Visitor to make Outputter and SymTable independent
This is done so that the symbol table can be populated once and the output can
be of any form. Constraining the traversal to be done using a Visitor, which
contains both SymTable and Outputter results in less flexible outputs. This is
useful when outputting in different forms such as Verilog and Dot language.