Initial implementation of L5X parser
The syntax of the hshg files are slightly modified so that it can have a
leading integer followed by a colon (":") that indicates the rung count. This
increases readability when looking at the file generated by the L5X parser,
which is then fed into the bison parser. With this change, the definition of
T_ADDRESS is changed to exclude numbers-only tokens. This allows
distinguishing constant values (like 42) from normal addresses (like
Local.4:0).